Fold Daemon type into Process type
The initial pass of connecting ConfigManager / Process type to Daemon
just added Daemon as an anonymous field in Process and removed dup
fields.  However, yaml/json marshalling ignores anonymous fields.
Don't need the two types anyhow, so this simplifies things in general.

This change is a large diff, but for the most part is just a rename of
Daemon to Process (for types and variables)

Change-Id: I04e88e9bd3faf61bde1b6a4c9805e3eb679aca71