Remove Process.Detached option
The Detached option was more useful when we did the double-fork self-detaching.
Since removing the double-fork, this option is less useful and will likely lead
to more confusion than convenience, where convenience was just gonit writing the
pid file.

With this option removed, processes are required to write their own pid file.

Change-Id: Id28025bfe3c58ed376f352f0d2f258b2792daa9a